requirements.txt
adal==1.2.2
azure-common==1.1.25
azure-datalake-store==0.0.48
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==9.0.0
azure-nspkg==3.0.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
idna==2.9
isodate==0.6.0
msrest==0.6.13
msrestazure==0.6.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
